REFACTOR: combine core documents files (#13733)
