+++ libs/core/langchain_core/documents/base.py
