+++ libs/core/langchain_core/documents/transformers.py
