+++ libs/core/tests/unit_tests/documents/test_imports.py
@@ -1,0 +1,7 @@
+from langchain_core.documents import __all__
+
+EXPECTED_ALL = ["Document", "BaseDocumentTransformer"]
+
+
+def test_all_imports() -> None:
+    assert set(__all__) == set(EXPECTED_ALL)
