+++ libs/langchain/langchain/document_transformers/beautiful_soup_transformer.py
@@ -1,7 +1,6 @@
 from typing import Any, Iterator, List, Sequence, cast
 
-from langchain_core.document_transformers import BaseDocumentTransformer
-from langchain_core.documents import Document
+from langchain_core.documents import BaseDocumentTransformer, Document
 
 
 class BeautifulSoupTransformer(BaseDocumentTransformer):
