+++ libs/langchain/langchain/document_transformers/doctran_text_translate.py
@@ -1,7 +1,6 @@
 from typing import Any, Optional, Sequence
 
-from langchain_core.document_transformers import BaseDocumentTransformer
-from langchain_core.documents import Document
+from langchain_core.documents import BaseDocumentTransformer, Document
 
 from langchain.utils import get_from_env
