+++ libs/langchain/langchain/document_transformers/embeddings_redundant_filter.py
@@ -2,8 +2,7 @@
 from typing import Any, Callable, List, Sequence
 
 import numpy as np
-from langchain_core.document_transformers import BaseDocumentTransformer
-from langchain_core.documents import Document
+from langchain_core.documents import BaseDocumentTransformer, Document
 from langchain_core.embeddings import Embeddings
 from langchain_core.pydantic_v1 import BaseModel, Field
