+++ libs/langchain/langchain/document_transformers/html2text.py
@@ -1,7 +1,6 @@
 from typing import Any, Sequence
 
-from langchain_core.document_transformers import BaseDocumentTransformer
-from langchain_core.documents import Document
+from langchain_core.documents import BaseDocumentTransformer, Document
 
 
 class Html2TextTransformer(BaseDocumentTransformer):
