+++ libs/langchain/langchain/document_transformers/long_context_reorder.py
@@ -1,8 +1,7 @@
 """Reorder documents"""
 from typing import Any, List, Sequence
 
-from langchain_core.document_transformers import BaseDocumentTransformer
-from langchain_core.documents import Document
+from langchain_core.documents import BaseDocumentTransformer, Document
 from langchain_core.pydantic_v1 import BaseModel
