+++ libs/langchain/langchain/retrievers/document_compressors/base.py
@@ -3,8 +3,7 @@ from abc import ABC, abstractmethod
 from inspect import signature
 from typing import List, Optional, Sequence, Union
 
-from langchain_core.document_transformers import BaseDocumentTransformer
-from langchain_core.documents import Document
+from langchain_core.documents import BaseDocumentTransformer, Document
 from langchain_core.pydantic_v1 import BaseModel
 
 from langchain.callbacks.manager import Callbacks
