+++ libs/langchain/langchain/schema/document.py
@@ -1,4 +1,3 @@
-from langchain_core.document_transformers import BaseDocumentTransformer
-from langchain_core.documents import Document
+from langchain_core.documents import BaseDocumentTransformer, Document
 
 __all__ = ["Document", "BaseDocumentTransformer"]
